Split navigation module into keyboard and slider navigation
